ocr-service/runpod/vlm/requirements.txt
# RunPod VLM Service Dependencies (VLM Only — no OCR engines)
# ============================================================
# NOTE: torch and torchvision are provided by the base image.
# vLLM is installed separately in Dockerfile. Do NOT add here.
# ============================================================

# RunPod SDK
runpod

# Image Processing (for PDF page conversion)
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
pillow>=10.0.0

# PDF support (VLM can process PDF pages as images)
pdf2image>=1.16.0

# DOCX support
python-docx>=1.1.0

# HTTP requests (for vLLM API and image URL fetching)
requests>=2.31.0

# HuggingFace
huggingface-hub
safetensors
